+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Fixed `TokenInterface->getRoles()` deprecation using `TokenInterface->getRoleNames()` instead
 ### Added|Changed|Deprecated|Removed|Fixed|Security
-Nothing so far
+Nothing else so far
 
 ## 6.1.4 - 2021-06-23
 ### Added